+++ src/stores/feed/types.ts
@@ -4,6 +4,6 @@ export interface FeedStore {
   imagePreview: string;
   setContent: (content: string) => void;
   setImageFile: (file: File) => void;
-  loadFeedData: (feedId: string) => Promise<void>;
+  loadFeedData: (feedId: number) => Promise<void>;
   resetForm: () => void;
 }
